+++ aomc/aomc.go
@@ -3,9 +3,12 @@ Package aomc is for handling ApiOmat "customer" resources during design-time.
 
 For example, you can fetch all classes of an ApiOmat module.
 
-You can use the package like this:
-1. Create a new Client instance with aomc.NewDefaultClient()
-2. Call for example the client's GetClasses() method to fetch all classes of a given module
+Have a look at the GoDoc code examples to see how to use this package.
+
+All returned errors are github.com/pkg/errors errors and contain a stack trace.
+
+The returned structs differ from the JSON that the ApiOmat server returns,
+because some attribute names in the JSON are misleading, verbose or inconsistent.
 */
 package aomc
 
@@ -13,14 +16,18 @@ import (
 	"github.com/philippgille/apiomat-go/aomx"
 )
 
-// Client is a client for ApiOmat customer resources
+// Client is a client for ApiOmat customer resources.
+//
+// It has an embedded type, which is an implementation of the aomx.Client interface,
+// so you can use the interface's methods as well (like `Get(path string, params url.Values) (string, error)`).
 type Client struct {
 	// Embedded anonymous type
 	aomx.Client
 }
 
 // NewDefaultClient creates a new ApiOmat client with an underlying aomx.DefaultClient.
-// username, password and system may be empty.
+//
+// Username, password and system may be empty.
 // If username or password are empty, no HTTP Authorization header is set in the HTTP request.
 // If system is empty, no "X-Apiomat-System" header is set in the HTTP request, leading to "LIVE" being used as default by ApiOmat.
 func NewDefaultClient(baseURL string, username string, password string, system aomx.System) Client {
@@ -30,6 +37,8 @@ func NewDefaultClient(baseURL string, username string, password string, system a
 }
 
 // NewClient creates a new ApiOmat client with a given client that implements the aomx.Client interface.
+//
+// This is mainly useful for writing tests or when you need to customize the client's behavior.
 func NewClient(client aomx.Client) Client {
 	return Client{
 		Client: client,